pyproject.toml
[tool.poetry]
name = "semtle-web-dev"
version = "0.1.0"
description = ""
authors = ["taehwan heo <htth815@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
sqlalchemy = "^2.0.29"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-decouple = "^3.8"
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
pydantic = "^2.8.2"
bcrypt = "^4.2.0"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
